Qual: Fix some CommonIncoterm/DocGenerator/People typings (#30940)
* Fix extraparams property typing

* Fix argument typing

* Fix 'status' argument type in calls to dolGetButtonTitle

* Fix argument2 for price to 1 (was 2)

* Fix deliveryreceipt argument

* Fix PHPDoc dol_sort_array

* Fix range for int argument

* Fix arg type for str_pad and add some PHPDoc

* Fix natsort value for dol_sort_array

* Fix arguments on calls to dol_sort_array

* Reduce phpstan exceptions

* Qual: Add type hints for Stripe objects

# Qual: Add type hints for Stripe objects

Suggest the more specific implementation of the StripObject.

* Typing, initialisation

* Qual: Fix undeclared variables

* Update phan baseline

* Update baseline

* Qual: Add type hints for Stripe objects

# Qual: Add type hints for Stripe objects

Suggest the more specific implementation of the StripObject.

* Add PHPDoc for host class properties

* Ignore false positive (phan) for deprecated property

* Improve typing

* Update baseline

+++ htdocs/comm/action/list.php
@@ -737,7 +737,7 @@ $url = DOL_URL_ROOT.'/comm/action/card.php?action=create';
 $url .= '&apyear='.$tmpforcreatebutton['year'].'&apmonth='.$tmpforcreatebutton['mon'].'&apday='.$tmpforcreatebutton['mday'].'&aphour='.$tmpforcreatebutton['hours'].'&apmin='.$tmpforcreatebutton['minutes'];
 $url .= '&backtopage='.urlencode($_SERVER["PHP_SELF"].($newparam ? '?'.$newparam : ''));
 
-$newcardbutton = dolGetButtonTitle($langs->trans('AddAction'), '', 'fa fa-plus-circle', $url, '', $user->hasRight('agenda', 'myactions', 'create') || $user->hasRight('agenda', 'allactions', 'create'));
+$newcardbutton = dolGetButtonTitle($langs->trans('AddAction'), '', 'fa fa-plus-circle', $url, '', (int) ($user->hasRight('agenda', 'myactions', 'create') || $user->hasRight('agenda', 'allactions', 'create')));
 
 $param .= '&mode='.urlencode($mode);
